move instance v2p code into instance_network
